changes added

+++ src/main/java/globantWorkshop/models/dao/UserDao.java
@@ -63,7 +63,7 @@ public class UserDao {
   }
 
 
-  public User getById(long id) {
+  public User getById(int id) {
     return entityManager.find(User.class, id);
   }
